Android App Making Successful Retrofit Call to ShopTrawler REST
REST is now providing an endpoint that gathers all the information needed for the RetailerBlock. Now there is a model created based off this in Android for handling the Deserialization as part of the asynchronous Retrofit call using GSON.

Logger was key here for finding errors in getting the Response back (a lot of them surrounded Timestamps (java.sql.Timestamp), will need to look into this more for the POST methods that will need to be sent from the Mobile Application to the REST.

The REST handles checking for if the Content and Loyalty Reward blocks are the correct ones for the current date, and return them.